updated expiry to match protobufs

+++ contracts/hts-precompile/ExpiryHelper.sol
@@ -9,13 +9,13 @@ contract ExpiryHelper is FeeHelper {
 
     function createAutoRenewExpiry(
         address autoRenewAccount,
-        int32 autoRenewPeriod
+        int64 autoRenewPeriod
     ) internal view returns (IHederaTokenService.Expiry memory expiry) {
         expiry.autoRenewAccount = autoRenewAccount;
         expiry.autoRenewPeriod = autoRenewPeriod;
     }
 
-    function createSecondExpiry(int32 second) internal view returns (IHederaTokenService.Expiry memory expiry) {
+    function createSecondExpiry(int64 second) internal view returns (IHederaTokenService.Expiry memory expiry) {
         expiry.second = second;
     }
 }
