+++ contracts/hts-precompile/IHederaTokenService.sol
@@ -69,14 +69,14 @@ interface IHederaTokenService {
     struct Expiry {
         // The epoch second at which the token should expire; if an auto-renew account and period are
         // specified, this is coerced to the current epoch second plus the autoRenewPeriod
-        int32 second;
+        int64 second;
 
         // ID of an account which will be automatically charged to renew the token's expiration, at
         // autoRenewPeriod interval, expressed as a solidity address
         address autoRenewAccount;
 
         // The interval at which the auto-renew account will be charged to extend the token's expiry
-        int32 autoRenewPeriod;
+        int64 autoRenewPeriod;
     }
 
     /// A Key can be a public key from either the Ed25519 or ECDSA(secp256k1) signature schemes, where
@@ -741,7 +741,7 @@ interface IHederaTokenService {
     function wipeTokenAccount(
         address token,
         address account,
-        int32 amount
+        int64 amount
     ) external returns (int64 responseCode);
 
     /// Operation to wipe non fungible tokens from account
